+++ CHANGELOG.md
@@ -638,6 +638,7 @@
 - `workflows`:
   - [v0.1.0](services/workflows/CHANGELOG.md#v010)
     - **New**: API for STACKIT Workflows
+    - [Usage example](https://github.com/stackitcloud/stackit-sdk-go/tree/main/examples/workflows)
 
 ## Release (2026-04-07)
 - `alb`: [v0.13.1](services/alb/CHANGELOG.md#v0131)